tools/illustrations/requirements.txt
google-genai>=1.0
Pillow>=10.0
